Playing around.

+++ simple_algorithms/binary_search.h
@@ -8,10 +8,10 @@
 #include <vector>
 
 template<typename T>
-int binary_search(const std::vector<T> &input, int left, int right, const T &target) requires std::is_integral_v<T>
+long long binary_search(const std::vector<T> &input, long long left, long long right, const T &target) requires std::is_integral_v<T>
 {
 	while (left <= right) {
-		int mid = (left + right) / 2;
+		long long mid = (left + right) / 2;
 		if (input[mid] == target)
 			return mid;
 		if (input[mid] < target)
@@ -24,8 +24,8 @@ int binary_search(const std::vector<T> &input, int left, int right, const T &tar
 
 template<typename T>
 int parallel_binary_search(const std::vector<T> &input,
-						   int left,
-						   int right,
+						   long long left,
+						   long long right,
 						   const T &target,
 						   int number_of_threads) requires std::is_integral_v<T>
 {
@@ -34,8 +34,8 @@ int parallel_binary_search(const std::vector<T> &input,
 	std::vector<T> results(number_of_threads, {});
 	for (int i{}; i < number_of_threads; ++i) {
 		threads.emplace_back([&, i]() {
-			int start = left + i * ((right - left+ 1) / number_of_threads);
-			int end = i == number_of_threads - 1 ? right : start + ((right - left + 1) / number_of_threads) - 1;
+			long long start = left + i * ((right - left+ 1) / number_of_threads);
+			long long end = i == number_of_threads - 1 ? right : start + ((right - left + 1) / number_of_threads) - 1;
 			results[i] = binary_search(input, start, end, target);
 		});
